chore(metadata): consolidate utils and update logs
