+++ src/utils/__tests__/utf8Truncate.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { utf8Truncate } from '../utf8Truncate';
+import { utf8Truncate } from '../metadata';
 
 describe('utf8Truncate', () => {
   it('keeps valid emoji', () => {
@@ -16,6 +16,13 @@ describe('utf8Truncate', () => {
   it('limits to max bytes', () => {
     const big = 'a'.repeat(300);
     const result = utf8Truncate(big, 200);
+    const encoder = new TextEncoder();
+    expect(encoder.encode(result).length).toBeLessThanOrEqual(200);
+  });
+
+  it('handles long emoji string', () => {
+    const emoji = '🙂'.repeat(120);
+    const result = utf8Truncate(emoji, 200);
     const encoder = new TextEncoder();
     expect(encoder.encode(result).length).toBeLessThanOrEqual(200);
   });
