+++ src/utils/metadata/historySlice.ts
@@ -1,0 +1,3 @@
+export function historySlice<T>(messages: T[], limit = 40): T[] {
+  return messages.slice(-limit);
+}
